Moved fonAwesome Componenet in components folder
